Add Java basic class structure (Refactor at start of the Class names to Upper Case)
