+++ src/main/java/me/sebasorovaa/BungeePlayerSpoof/commands/AddPlayer.java
@@ -1,12 +1,12 @@
-package me.sebasorovaa.BungeePlayerSpoof;
+package me.sebasorovaa.BungeePlayerSpoof.commands;
 
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.plugin.Command;
 import net.md_5.bungee.api.ChatColor;
 import static me.sebasorovaa.BungeePlayerSpoof.Main.fakepla;
 
-public class addPlayer extends Command {
-    public addPlayer() {
+public class AddPlayer extends Command {
+    public AddPlayer() {
         super("addPlayer");
     }
     public void execute(CommandSender sender, String[] args) {
